+++ packages/papupata/src/main/responder.ts
@@ -19,7 +19,7 @@ import {
 } from './responderTypes'
 import runExpressMiddleware from './runExpressMiddleware'
 import { TypedQueryToTypes } from './TypedQueryToTypes'
-import { ActualTypeMap, Method, StringTupleElementTypes } from './types'
+import { Method } from './types'
 import {
   extractHardCodedParameters,
   getQueryWithHardCodedParameters,
@@ -30,7 +30,7 @@ import { paramMatchers } from './utils/paramMatchers'
 import { runHandlerChain } from './utils/runHandlerChain'
 
 export function responder<
-  ParamsType extends readonly string[],
+  ParamsType extends TypedQueryType,
   QueryType extends TypedQueryType,
   OptionalQueryType extends TypedQueryType,
   BoolQueryType extends TypedQueryType,
@@ -52,20 +52,6 @@ export function responder<
   routeOptions: RouteOptions,
   papupataOptions: PapupataRouteOptions
 ) {
-  /*type CallArgsWithoutBody = ActualTypeMap<StringTupleElementTypes<ParamsType>, string> &
-    ActualTypeMap<StringTupleElementTypes<QueryType>, string> &
-    ActualOptionalTypeMap<StringTupleElementTypes<OptionalQueryType>, string> &
-    ActualTypeMap<StringTupleElementTypes<BoolQueryType>, boolean>*/
-  //type CallArgs = BodyInputType & CallArgsWithoutBody
-
-  /*type CallArgParam = {} extends CallArgs
-    ? [] | [CallArgs] | [CallArgs, RequestOptions]
-    :
-        | [CallArgs]
-        | [CallArgs, RequestOptions]
-        | [BodyInputType, CallArgsWithoutBody]
-        | [BodyInputType, CallArgsWithoutBody, RequestOptions]*/
-
   return {
     response<ResponseType, ResponseTypeOnServer = ResponseType>(
       mapper?: (payload: ResponseTypeOnServer) => ResponseType | Promise<ResponseType>
@@ -120,7 +106,7 @@ export function responder<
 
         const requestOptions = separateBody && hasOtherArgs ? argsArr[2] : (argsArr[1] as any)
 
-        const reqParams = pick(args, params),
+        const reqParams = pick(args, Object.keys(params)),
           reqQuery = getQueryWithHardCodedParameters(hardCodedParameters, {
             ...pick(args, Object.keys(query)),
             ...pick(args, Object.keys(optionalQuery)),
@@ -129,7 +115,12 @@ export function responder<
           reqBody = separateBody
             ? argsArr[0]
             : makeUndefinedIfEmpty(
-                omit(args, [...params, ...Object.keys(query), ...Object.keys(boolQuery), ...Object.keys(optionalQuery)])
+                omit(args, [
+                  ...Object.keys(params),
+                  ...Object.keys(query),
+                  ...Object.keys(boolQuery),
+                  ...Object.keys(optionalQuery),
+                ])
               )
 
         if (mockImpl) {
@@ -162,7 +153,12 @@ export function responder<
 
       function isValidAsNonBodyRequestData(obj: any) {
         if (typeof obj !== 'object') return false
-        const validKeys = [...Object.keys(query), ...Object.keys(optionalQuery), ...Object.keys(boolQuery), ...params]
+        const validKeys = [
+          ...Object.keys(query),
+          ...Object.keys(optionalQuery),
+          ...Object.keys(boolQuery),
+          ...Object.keys(params),
+        ]
         return Object.keys(obj).every((key) => validKeys.includes(key))
       }
 
@@ -259,12 +255,15 @@ export function responder<
           }
         }
         try {
+          const convertedParams = handleQueryParameterTypes(req.params, params, Mode.REQUIRED)
           const queryConversion1 = handleQueryParameterTypes(req.query, query, Mode.REQUIRED)
           const queryConversion2 = handleQueryParameterTypes(queryConversion1, boolQuery, Mode.LEGACY_BOOL)
           const convertedQuery = handleQueryParameterTypes(queryConversion2, optionalQuery, Mode.OPTIONAL)
 
-          const originalQuery = req.query
+          const originalQuery = req.query,
+            originalParams = req.params
           req.query = convertedQuery
+          req.params = convertedParams
           let value: any
           try {
             value = await runHandlerChain(
@@ -279,6 +278,7 @@ export function responder<
             )
           } finally {
             req.query = originalQuery
+            req.params = originalParams
           }
           if (value === skipHandlingRoute) {
             return next()
@@ -327,10 +327,8 @@ export function responder<
 
       function getURL(
         pathParamsAndQueryParams:
-          | ActualTypeMap<StringTupleElementTypes<ParamsType>, string>
-          | (ActualTypeMap<StringTupleElementTypes<ParamsType>, string> &
-              TypedQueryToTypes<QueryType & BoolQueryType> &
-              Partial<TypedQueryToTypes<OptionalQueryType>>)
+          | TypedQueryToTypes<ParamsType>
+          | (TypedQueryToTypes<QueryType & BoolQueryType & ParamsType> & Partial<TypedQueryToTypes<OptionalQueryType>>)
       ) {
         const config = parent.getConfig()
         if (!config) throw new Error('Papupata not configured')
@@ -347,14 +345,17 @@ export function responder<
       call.CallArgsType = null as any
 
       return call
-      function applyPathParams(reqParams: ActualTypeMap<StringTupleElementTypes<ParamsType>, string>) {
+      function applyPathParams(reqParams: TypedQueryToTypes<ParamsType>) {
         const pathWithParams = paramMatchers(params).reduce((currPath, { matcher, name }) => {
           return currPath.replace(matcher, (_, before, after) => {
             return `${before}${encodeURIComponent((reqParams as any)[name])}${after}`
           })
         }, path)
 
-        const queryParams = getQueryWithHardCodedParameters(hardCodedParameters, omit(reqParams, [...params]) as any)
+        const queryParams = getQueryWithHardCodedParameters(
+          hardCodedParameters,
+          omit(reqParams, [...Object.keys(params)]) as any
+        )
         if (Object.keys(queryParams).length) {
           return pathWithParams + '?' + qs.stringify(queryParams)
         } else {